+++ tests/test_plugin_semver2.py
@@ -60,7 +60,7 @@ def test_tag_pyproject(tmpdir, ctlr):
 
     plugin.tag(version="2.0.0", repo="dummy_repo", prerelease="rc")
 
-    with open(pyproject_path, "r") as f:
+    with open(pyproject_path) as f:
         pyproject = tomlkit.load(f)
     assert pyproject["tool"]["poetry"]["version"] == "2.0.0-rc.1"